Universal and Startup Systems Implemented
A system which contains a class which has things that are required in every single class in the entire project. Additionally, a separate class has been added for the sole purpose of telling the menu to start so that the menu class can also extend the universal class even though it is the only class with a main string args in it. This is meant to prevent any problems which could arise in the future.